requirements.txt
anyio==4.12.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
duckduckgo_search==8.1.1
fake-useragent==2.2.0
googlesearch-python==1.3.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
playwright==1.57.0
playwright-stealth==2.0.0
primp==0.15.0
psycopg2-binary==2.9.11
pycparser==2.23
pyee==13.0.0
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.1
requests==2.32.5
socksio==1.0.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.6.1
python-dotenv